refactor: remove prepare_report_mapping_context indirection from mapping.py
_build_report_template_data() now uses prepared.mapping_context directly
instead of calling prepare_report_mapping_context(prepared), removing an
unnecessary indirection layer. The function is kept in report_context.py
for test convenience. Test imports updated to use the canonical source.

Closes #1398
